AORTA-17 CLI command for report generation : Implementation of pipeline (#70)
* AORTA-17 CLI command for report generation : Implementation of pipeline

* AORTA-17 CLI command for report generation : refactored cli (#71)


---------
