+++ src/aorta/report/pipelines/__init__.py
@@ -1,0 +1,14 @@
+"""Pipeline orchestrators for multi-step analysis workflows."""
+
+from .summary_pipeline import run_summary_pipeline, SummaryPipelineConfig, PipelineResult
+from .gemm_pipeline import run_gemm_pipeline, GemmPipelineConfig, GemmPipelineResult
+
+__all__ = [
+    "run_summary_pipeline",
+    "SummaryPipelineConfig",
+    "PipelineResult",
+    "run_gemm_pipeline",
+    "GemmPipelineConfig",
+    "GemmPipelineResult",
+]
+
